requirements.txt
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0